refactor: Move types to their own file for better reusability
