+++ src/components/TextBox/type.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export interface TextBoxProps {
+  children: ReactNode;
+}
