+++ src/pages/Login/type.ts
@@ -1,0 +1,8 @@
+import { KeyboardTypeOptions } from "react-native";
+
+export interface FormInputsType {
+  label: string;
+  name: string;
+  keyboard: KeyboardTypeOptions;
+  autoComplete: "email" | "password";
+}
